amdsmi: revert partition wrapper change and clarify docs
Revert the wrapper portion of f9efbd8629 ("doc cleanup") in
amdsmi_get_gpu_accelerator_partition_profile_config(). That commit
changed the wrapper's return shape without updating amdsmi_commands.py,
which caused `amd-smi partition -a` to raise
`IndexError: list index out of range`.

Changes:
- py-interface: restore original wrapper behavior so the inner resource
  loop iterates `config.num_resource_profiles` and the top-level
  `resource_profiles` aliases the last profile's `resources`.
- docs/reference/amdsmi-py-api.md: rewrite field descriptions for
  amdsmi_get_gpu_accelerator_partition_profile_config to match what the
  wrapper actually returns.
- docs/conceptual/partition.md: add a comparison note explaining why
  MI300X supports SPX/DPX/QPX/CPX while MI300A supports SPX/DPX/TPX/CPX
  (XCD-count divisibility), and wrap the orphan NPS8 paragraph in a
  {note} block.
- example/amd_smi_partition_example.cc: use uint32_t for GPU indexes in
  PrimaryGpuList to match the amdsmi C API integer width, keeping the
  example warning-clean under -Wconversion / -Wshorten-64-to-32 in case
  those flags are later enabled for the example targets.

+++ projects/amdsmi/example/amd_smi_partition_example.cc
@@ -232,7 +232,7 @@ class AmdsmiSession {
 // partition config; sub-partition handles return NOT_SUPPORTED. Calling set on
 // a sub-partition handle is a no-op (NOT_SUPPORTED), so filter them out first.
 // Each entry is {amd-smi GPU index, handle}.
-using PrimaryGpuList = std::vector<std::pair<size_t, amdsmi_processor_handle>>;
+using PrimaryGpuList = std::vector<std::pair<uint32_t, amdsmi_processor_handle>>;
 
 [[nodiscard]] static PrimaryGpuList get_primary_gpu_handles(
     const std::vector<amdsmi_processor_handle>& all_handles) {
@@ -241,7 +241,7 @@ using PrimaryGpuList = std::vector<std::pair<size_t, amdsmi_processor_handle>>;
     amdsmi_accelerator_partition_profile_config_t cfg{};
     if (amdsmi_get_gpu_accelerator_partition_profile_config(all_handles[i], &cfg) ==
         AMDSMI_STATUS_SUCCESS)
-      primary.emplace_back(i, all_handles[i]);
+      primary.emplace_back(static_cast<uint32_t>(i), all_handles[i]);
   }
   return primary;
 }
